feat: replace rainbow with onboard
