+++ packages/frontend/package.json
@@ -33,6 +33,7 @@
         "@material-ui/icons": "^4.11.2",
         "@material-ui/lab": "^4.0.0-alpha.60",
         "@material-ui/pickers": "^3.3.10",
+        "@rainbow-me/rainbowkit": "^0.8.1",
         "@testing-library/react-hooks": "^8.0.0",
         "@thanpolas/univ3prices": "^3.0.2",
         "@uniswap/sdk-core": "^3.0.1",
@@ -46,6 +47,7 @@
         "clsx": "^1.1.1",
         "copy-to-clipboard": "^3.3.1",
         "date-fns": "^2.29.3",
+        "ethers": "^5.7.2",
         "fathom-client": "^3.2.0",
         "firebase": "^8.8.0",
         "framer-motion": "^5.3.3",
@@ -71,6 +73,7 @@
         "set-interval-async": "^2.0.3",
         "subscriptions-transport-ws": "^0.11.0",
         "use-debounce": "^9.0.2",
+        "wagmi": "^0.5.5",
         "web3": "^1.5.2"
     },
     "devDependencies": {
